+++ src/db.js
@@ -43,7 +43,7 @@ var sequelize = new Sequelize(dbConfig.database, dbConfig.user, dbConfig.passwor
 	dialectOptions,
 	timeZone       : config.timeZone,
 	logging        : require('debug')('app:db:query'),
- 	logging				 : console.log,
+ 	//logging				 : console.log,
 	typeValidation : true,
 
 	define: {