+++ src/lib/sequelize-authorization/mixins/authorize-data.js
@@ -16,8 +16,7 @@ module.exports = function authorizeData(data, action, user, self, project) {
       userId = self.id
     }
 
-    // TODO: dit is een check op jezelf, nu kan de argument:view check uit de routes
-    if (!self.can(action, user))  throw 'cannot';
+    if (!self.can(action, user)) throw 'cannot';
 
     let keys = Object.keys( data );